requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
django-environ==0.11.2
django-extensions==3.2.3
psycopg2-binary==2.9.9
Pillow==10.3.0
python-slugify==8.0.4
drf-spectacular==0.27.2
django-ratelimit==4.1.0
whitenoise==6.7.0
gunicorn==22.0.0
celery==5.4.0
django-celery-beat==2.6.0
redis==5.0.4
